test: fix `C4244` warning in the result from function getting testing

+++ test/result_test.cpp
@@ -86,7 +86,7 @@ TEST_CASE("Result rewriting") {
 }
 
 TEST_CASE("Result from function getting") {
-  const auto square_root = [](int value) -> result::Result<int> {
+  const auto square_root = [](double value) -> result::Result<double> {
     if (value < 0) return error::Error("value must be positive");
     return std::sqrt(value);
   };
